refactor: rename plain/common-types to actually describe what the file contains
