+++ lib/contentful-management.ts
@@ -11,7 +11,7 @@ import type { AdapterParams } from './create-adapter'
 import { createAdapter } from './create-adapter'
 import type { ClientAPI } from './create-contentful-api'
 import createContentfulApi from './create-contentful-api'
-import type { PlainClientAPI } from './plain/common-types'
+import type { PlainClientAPI } from './plain/plain-client-types'
 import type { DefaultParams } from './plain/plain-client'
 import { createPlainClient } from './plain/plain-client'
 import * as editorInterfaceDefaults from './constants/editor-interface-defaults/index'
@@ -20,7 +20,7 @@ export type { ClientAPI } from './create-contentful-api'
 export { asIterator } from './plain/as-iterator'
 export { fetchAll } from './plain/pagination-helper'
 export { isDraft, isPublished, isUpdated } from './plain/checks'
-export type { PlainClientAPI } from './plain/common-types'
+export type { PlainClientAPI } from './plain/plain-client-types'
 export { createClient }
 export { RestAdapter } from './adapters/REST/rest-adapter'
 export type { RestAdapterParams } from './adapters/REST/rest-adapter'
