+++ lib/plain/plain-client.ts
@@ -1,6 +1,6 @@
 import type { GetContentTypeParams, GetSpaceEnvironmentParams, MakeRequest } from '../common-types'
 import { omitAndDeleteField } from '../methods/content-type'
-import type { PlainClientAPI } from './common-types'
+import type { PlainClientAPI } from './plain-client-types'
 import type { DefaultParams } from './wrappers/wrap'
 import { wrap } from './wrappers/wrap'
